combine two files to one
